+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0
+
+- Added min (Radius.zero) & max (Size/2) radius to `BoxShape.rectangle`.
+- Resolved `BoxShape.circle` conflicts with borderRadius.
+
 ## 0.1.0
 
-* Initial release.
+- Initial release.